+++ tests/urls.py
@@ -1,7 +1,10 @@
 from django.urls import path
 
-from tests.views import ConcretAtomicOperationView
+from tests.views import BulkAtomicOperationView, ConcretAtomicOperationView
+
 
 urlpatterns = [
-    path("", ConcretAtomicOperationView.as_view())
+    path("", ConcretAtomicOperationView.as_view()),
+    path("bulk", BulkAtomicOperationView.as_view())
+
 ]